Remove usage of reflect on JSON tags for getting metric names and push datapoint building to be verbose so we can easily add metrics units.
