+++ collectors/mesos/agent/state.go
@@ -1,0 +1,65 @@
+package agent
+
+import (
+	"net"
+	"net/url"
+	"strconv"
+
+	"github.com/dcos/dcos-metrics/util/http/client"
+)
+
+// agentState defines the structure of the response expected from Mesos
+// *for all cluster state* when polling the /state endpoint.
+// Specifically, this struct exists for the following purposes:
+//
+//   * collect labels from individual containers (executors) since labels are
+//     NOT available via the /containers endpoint in v1.
+//   * map framework IDs to a human-readable name
+//
+// For more information, see the upstream protobuf in Mesos v1:
+//
+//   * Framework info: https://github.com/apache/mesos/blob/1.0.1/include/mesos/v1/mesos.proto#L207-L307
+//   * Executor info:  https://github.com/apache/mesos/blob/1.0.1/include/mesos/v1/mesos.proto#L474-L522
+//
+type agentState struct {
+	ID         string          `json:"id"`
+	Hostname   string          `json:"hostname"`
+	Frameworks []frameworkInfo `json:"frameworks"`
+}
+
+type frameworkInfo struct {
+	ID        string         `json:"id"`
+	Name      string         `json:"name"`
+	Principal string         `json:"principal,omitempty"`
+	Role      string         `json:"role"`
+	Executors []executorInfo `json:"executors,omitempty"`
+}
+
+type executorInfo struct {
+	ID        string           `json:"id"`
+	Name      string           `json:"name"`
+	Container string           `json:"container"`
+	Labels    []executorLabels `json:"labels,omitempty"` // labels are optional
+}
+
+type executorLabels struct {
+	Key   string `json:"key"`
+	Value string `json:"value"`
+}
+
+// getAgentState fetches the state JSON from the Mesos agent, which contains
+// info such as framework names and IDs, the current leader, config flags,
+// container (executor) labels, and more.
+func (c *Collector) getAgentState() error {
+	c.agentState = agentState{}
+
+	u := url.URL{
+		Scheme: c.RequestProtocol,
+		Host:   net.JoinHostPort(c.nodeInfo.IPAddress, strconv.Itoa(c.Port)),
+		Path:   "/state",
+	}
+
+	c.HTTPClient.Timeout = HTTPTIMEOUT
+
+	return client.Fetch(c.HTTPClient, u, &c.agentState)
+}
